SNOW-1631831: JDBC - handle TCP connection closure (traffic dependency request) (#1872)
